src/main/java/com/interview/rewards/calculator/service/TransactionService.java
package com.interview.rewards.calculator.service;

import com.interview.rewards.calculator.model.Transaction;
import com.interview.rewards.calculator.repository.TransactionRepository;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

/**

 Service class for performing CRUD operations on Transaction entity.
 */
@Service
public class TransactionService {
    private static final Logger LOGGER = LoggerFactory.getLogger(CustomerService.class);

    private final TransactionRepository transactionRepository;

    public TransactionService(TransactionRepository transactionRepository) {
        this.transactionRepository = transactionRepository;
    }

    public Transaction save(Transaction transaction) {
        return transactionRepository.save(transaction);
    }

    public List<Transaction> saveAll(List<Transaction> transactions) {
        return transactionRepository.saveAll(transactions);
    }

    public List<Transaction> findAll() {
        return transactionRepository.findAll();
    }

    public Optional<Transaction> findById(Long id) {
        return transactionRepository.findById(id);
    }

    public void deleteById(Long id) {
        transactionRepository.deleteById(id);
    }
}